src/deployment.py
```python
# Importing necessary libraries
import tensorflow as tf
from onnxruntime import InferenceSession
from tritonclient.grpc import service_pb2_grpc, service_pb2, grpc_service
import grpc

class Deployment:
    def __init__(self, model):
        self.model = model

    def tensorflow_serving(self, model_dir, model_version):
        """
        Function to serve a TensorFlow model
        """
        # Load the model
        model = tf.saved_model.load(model_dir)

        # Serve the model
        tf.saved_model.save(model, f"{model_dir}/{model_version}")

    def onnx_runtime(self, model_path):
        """
        Function to serve an ONNX model
        """
        # Load the model
        sess = InferenceSession(model_path)

        # Serve the model
        input_name = sess.get_inputs()[0].name
        label_name = sess.get_outputs()[0].name
        return sess, input_name, label_name

    def nvidia_triton(self, model_repository, model_name, model_version):
        """
        Function to serve a model using Nvidia Triton Inference Server
        """
        # Create a gRPC stub
        channel = grpc.insecure_channel('localhost:8001')
        grpc_stub = service_pb2_grpc.GRPCInferenceServiceStub(channel)

        # Create a request for model status
        request = service_pb2.StatusRequest(model_name=model_name)

        # Get model status
        response = grpc_stub.Status(request)

        # Check the model version status
        version_status = response.model_status[model_name].version_status[model_version]
        if version_status.ready_state == service_pb2.ModelReadyState.READY:
            print(f"Model {model_name} (version {model_version}) is ready to serve requests.")
        else:
            print(f"Model {model_name} (version {model_version}) is not ready to serve requests.")

    def model_deployment_best_practices(self):
        """
        Function to outline model deployment best practices
        """
        print("""
        1. Test the model thoroughly before deployment.
        2. Monitor the model's performance continuously after deployment.
        3. Plan for model updates and rollbacks.
        4. Ensure the security and privacy of the model and the data it uses.
        5. Use a model serving solution that matches your needs (scalability, latency, etc.).
        """)

    def multi_model_serving(self):
        """
        Function to outline strategies for serving multiple models
        """
        print("""
        1. Use a model serving platform that supports multi-model serving.
        2. Consider the resource requirements of each model.
        3. Use a load balancer to distribute requests among multiple models.
        4. Monitor the performance of each model individually.
        """)
```
