tests/test_deployment.py
```python
# Importing necessary libraries
import unittest
import tensorflow as tf
from onnxruntime import InferenceSession
from tritonclient.grpc import service_pb2_grpc, service_pb2, grpc_service
import grpc

from src.deployment import Deployment

class TestDeployment(unittest.TestCase):
    def setUp(self):
        self.deployment = Deployment('test_model')

    def test_tensorflow_serving(self):
        # Mocking the model directory and version
        model_dir = 'test_dir'
        model_version = '1'

        # Call the tensorflow_serving method
        self.deployment.tensorflow_serving(model_dir, model_version)

        # Check if the model is saved in the correct directory
        self.assertTrue(tf.saved_model.contains_saved_model(f"{model_dir}/{model_version}"))

    def test_onnx_runtime(self):
        # Mocking the model path
        model_path = 'test_model.onnx'

        # Call the onnx_runtime method
        sess, input_name, label_name = self.deployment.onnx_runtime(model_path)

        # Check if the session is an instance of InferenceSession
        self.assertIsInstance(sess, InferenceSession)

    def test_nvidia_triton(self):
        # Mocking the model repository, name, and version
        model_repository = 'test_repository'
        model_name = 'test_model'
        model_version = '1'

        # Call the nvidia_triton method
        self.deployment.nvidia_triton(model_repository, model_name, model_version)

        # Check if the model is ready to serve requests
        channel = grpc.insecure_channel('localhost:8001')
        grpc_stub = service_pb2_grpc.GRPCInferenceServiceStub(channel)
        request = service_pb2.StatusRequest(model_name=model_name)
        response = grpc_stub.Status(request)
        version_status = response.model_status[model_name].version_status[model_version]
        self.assertEqual(version_status.ready_state, service_pb2.ModelReadyState.READY)

    def test_model_deployment_best_practices(self):
        # Call the model_deployment_best_practices method
        self.deployment.model_deployment_best_practices()

        # Check if the method prints the correct output
        self.assertPrints("""
        1. Test the model thoroughly before deployment.
        2. Monitor the model's performance continuously after deployment.
        3. Plan for model updates and rollbacks.
        4. Ensure the security and privacy of the model and the data it uses.
        5. Use a model serving solution that matches your needs (scalability, latency, etc.).
        """)

    def test_multi_model_serving(self):
        # Call the multi_model_serving method
        self.deployment.multi_model_serving()

        # Check if the method prints the correct output
        self.assertPrints("""
        1. Use a model serving platform that supports multi-model serving.
        2. Consider the resource requirements of each model.
        3. Use a load balancer to distribute requests among multiple models.
        4. Monitor the performance of each model individually.
        """)

if __name__ == '__main__':
    unittest.main()
```
